Implemented suggestion to move tests to NSubstitute.Acceptance.Specs
See https://github.com/nsubstitute/NSubstitute/pull/808/files#r1599922691
