+++ tests/NSubstitute.Acceptance.Specs/AsyncEventHandlers/ITestInterface.cs
@@ -1,8 +1,12 @@
-﻿namespace TestAsyncEventHandlersWithNSubstiute;
+﻿namespace NSubstitute.Acceptance.Specs.AsyncEventHandlers;
 
 public delegate Task TestEventHandler();
 
 public interface ITestInterface
 {
+    #region Events
+
     event TestEventHandler TestEvent;
+
+    #endregion
 }
