--- tests/Nsubstitute.TestAsyncEventHandlersWithNSubstiute/TestImplementation.cs
@@ -1,13 +1,0 @@
-﻿namespace TestAsyncEventHandlersWithNSubstiute
-{
-
-    public class TestImplementation : ITestInterface
-    {
-        public event TestEventHandler? TestEvent;
-
-        public async Task RaiseEventAsync()
-        {
-            if (TestEvent != null) await TestEvent();
-        }
-    }
-}
